+++ docs/reference/changelog.rst
@@ -15,7 +15,7 @@ Changelog
 
     For a complete list of commits, check out the `1.2.3`_ release on GitHub.
 
-7.0.1 (unreleased)
+7.1.1 (unreleased)
 ------------------
 
 Bug fixes
@@ -23,7 +23,28 @@ Bug fixes
 
 - ``remote-build`` no longer raises an error when dangling symlinks are present in the
   project repository.
+- Ignore .spread-reuse files when deciding if a source is outdated.
 
+7.1.0 (2026-07-07)
+------------------
+
+Services
+========
+
+- Add the metadata mediator API to the package service to only repack artifacts
+  when the lifecycle or metadata files changed.
+
+For a complete list of commits, check out the `7.1.0`_ release on GitHub.
+
+7.0.1 (2026-07-03)
+------------------
+
+Bug fixes
+=========
+
+- Pass proxy environment variables to the fetch-service.
+
+For a complete list of commits, check out the `7.0.1`_ release on GitHub.
 
 7.0.0 (2026-06-02)
 ------------------
@@ -1367,3 +1388,5 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _6.3.1: https://github.com/canonical/craft-application/releases/tag/6.3.1
 .. _6.4.0: https://github.com/canonical/craft-application/releases/tag/6.4.0
 .. _7.0.0: https://github.com/canonical/craft-application/releases/tag/7.0.0
+.. _7.0.1: https://github.com/canonical/craft-application/releases/tag/7.0.1
+.. _7.1.0: https://github.com/canonical/craft-application/releases/tag/7.1.0